Move VhostRedirectView to better place
